requirements.txt
Flask==2.3.3
groq==0.4.1
pywhatkit==5.4
mysql-connector-python==8.1.0
requests==2.31.0

